Modularização do menu
